Redo the board layout to be more compact.
IMPORTANT: as of this moment, this is only a proposal for
a new design, and the team has *not* determined whether
this is necessary.

This addresses some issues that I have had with the design
of the Controls Leader board:
 - It is too bulky, and would be difficult to mount
   where it would be out of the way.
 - It requires the driver to physically interact with it,
   which could put unecessary strain on the board and
   further restrict the areas of the car where it could
   be mounted.
 - All of the connectors are on the back of the board,
   making it difficult to interact with during testing.

Additionally, a strong effort has been made to maintain
software compatibility between this board and the now-functional
existing hardware.